Undo previous change to package.json
In package.json, the change was to use Chrome headless and get out of
watch mode (which was the default and isn't great for Travis-CI), but I
went about that in the wrong manner.

This commit makes a change in karma.conf.js with the same goals. It also
removes the use of single run since that is apparently deprecated.